requirements.txt
# AI Model SDKs
anthropic==0.34.0
openai==1.40.0
groq==0.11.0
google-generativeai==0.8.0
xai-sdk==1.5.0

# AI Orchestration
litellm==1.40.0

# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.8.0
pandas==2.2.0
numpy==1.26.0
pydantic==2.8.0

# UI/Display
rich==13.7.0
tqdm==4.66.0
termcolor==2.3.0
bokeh==3.4.0
matplotlib==3.8.0

# Trading/Backtesting
backtesting==0.3.3

# Text-to-Speech
pyttsx3==2.90

# Telegram
python-telegram-bot==20.0

# SSL/Networking
certifi==2024.8.30
urllib3==2.2.0
idna==3.7
charset-normalizer==3.3.0